+++ src/_mocks/TasteFinderMock.ts
@@ -58,6 +58,16 @@ export const TasteFinderMockQueryEspresso = {
     data: {
       tasteFinderRecommendation: {
         shopifyProductIds: ['7655228866776'],
+        recommendations: [
+          {
+            shopifyId: '7655228866776',
+            score: 0.9814814814814815,
+            acidity: 4,
+            bitterness: 1,
+            sweetness: 2,
+            body: 3,
+          },
+        ],
       },
     },
   },
@@ -80,6 +90,16 @@ export const TasteFinderMockQueryFilter = {
     data: {
       tasteFinderRecommendation: {
         shopifyProductIds: ['7655228866776'],
+        recommendations: [
+          {
+            shopifyId: '7655228866776',
+            score: 0.9814814814814815,
+            acidity: 4,
+            bitterness: 1,
+            sweetness: 2,
+            body: 3,
+          },
+        ],
       },
     },
   },