+++ src/views/Product/FeaturedProduct.tsx
@@ -32,8 +32,9 @@ import {
   TypographySize,
   TypographyType,
 } from 'src/components'
-import { getAPIProductId, getAromaKey } from 'src/utils'
+import { getAPIProductId, getAromaKey, useLocalStorage } from 'src/utils'
 
+import { TasteFinderField } from '../TasteFinder'
 import { getGuideImages, getTasteResults, TasteProfile } from '.'
 
 interface ProductMeta {
@@ -80,6 +81,8 @@ export const FeaturedProduct: React.FC = () => {
   const { id } = useParams()
   const { t } = useTranslation()
   const GET_PRODUCT = loader('src/graphql/queries/product.query.graphql')
+  const [tasteFinderDataJSON] = useLocalStorage('tasteFinder', '')
+  const tasteFinderData = tasteFinderDataJSON && JSON.parse(tasteFinderDataJSON)
 
   useEffect(() => {
     document.title = `${t('brand.name')} | ${t('pages.product.title')}`
@@ -95,12 +98,11 @@ export const FeaturedProduct: React.FC = () => {
 
   const { aroma, images, title, whyThisCoffee } = data?.product || {}
 
-  // TODO - this is hardcoded for now, as next: grab the stored taste profile data
   const tasteProfileResults: TasteProfile = {
-    acidity: 1,
-    bitterness: 2,
-    sweetness: 3,
-    body: 1,
+    acidity: tasteFinderDataJSON && tasteFinderData.find((el: TasteFinderField) => el.name === 'acidity').value,
+    bitterness: tasteFinderDataJSON && tasteFinderData.find((el: TasteFinderField) => el.name === 'bitterness').value,
+    sweetness: tasteFinderDataJSON && tasteFinderData.find((el: TasteFinderField) => el.name === 'sweetness').value,
+    body: tasteFinderDataJSON && tasteFinderData.find((el: TasteFinderField) => el.name === 'body').value,
   }
 
   if (error) {
@@ -111,6 +113,23 @@ export const FeaturedProduct: React.FC = () => {
     return (
       <div className="flex h-64 mb-32 justify-center items-center">
         <Loader />
+      </div>
+    )
+  }
+
+  const getMatchScore = () => {
+    const recommendations =
+      tasteFinderDataJSON && tasteFinderData.find((el: TasteFinderField) => el.name === 'recommendations')?.value
+    const score =
+      tasteFinderDataJSON && recommendations && recommendations[0]?.score
+        ? Math.round(recommendations[0]?.score * 100)
+        : 0
+
+    if (score === 0) return null
+
+    return (
+      <div className="mt-4">
+        <Tag type={TagType.TasteFinder} value={`${score}% Übereinstimmung`} />
       </div>
     )
   }
@@ -215,7 +234,7 @@ export const FeaturedProduct: React.FC = () => {
         <div className="grid grid-cols-2 gap-0 mt-8 md:gap-4">
           <Link to={`/product/${id}`} className="flex justify-center items-start md:row-start-1 md:row-end-3">
             <div className="flex aspect-1 justify-center items-center rounded-full bg-coreUI-background-images w-3/4 shrink-0 grow-0">
-              <img src={images.nodes[0].url} alt={title} className="w-full w-3/4 h-fit shrink-0 grow-0" />
+              <img src={images.nodes[0].url} alt={title} className="w-full w-3/4 h-auto shrink-0 grow-0" />
             </div>
           </Link>
           {data?.product && (
@@ -227,9 +246,7 @@ export const FeaturedProduct: React.FC = () => {
                 showFrom={true}
                 className="p-0"
               />
-              <div className="mt-4">
-                <Tag type={TagType.TasteFinder} value="98% Übereinstimmung" />
-              </div>
+              {getMatchScore()}
             </Link>
           )}
           <div className="col-span-2 md:col-start-2 md:col-span-1">