+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# [1.15.0](https://github.com/Famous-Roasters/frontend/compare/v1.14.0...v1.15.0) (2022-07-06)
+
+
+### Features
+
+* **taste-finder:** fix missing score bug ([b59526c](https://github.com/Famous-Roasters/frontend/commit/b59526c7d7a07e23d5c8639d2e79758556656afc))
+
+# [1.14.0](https://github.com/Famous-Roasters/frontend/compare/v1.13.1...v1.14.0) (2022-07-06)
+
+
+### Features
+
+* **taste finder:** dynamic taste match score ([8ec9fc1](https://github.com/Famous-Roasters/frontend/commit/8ec9fc186c775e8015d40b32e3878b5cfe42eb14))
+* **taste finder:** tests local storage ([339e6a2](https://github.com/Famous-Roasters/frontend/commit/339e6a29b654fb6fd554b19de5cc62bf3f0f1b67))
+
+## [1.13.1](https://github.com/Famous-Roasters/frontend/compare/v1.13.0...v1.13.1) (2022-07-06)
+
+
+### Bug Fixes
+
+* **product:** grind info popup typo ([a1d7a48](https://github.com/Famous-Roasters/frontend/commit/a1d7a48f93923282ade90745deb9c70ec6d9b9fb))
+* **taste finder:** fix image skew on result page ([be41dca](https://github.com/Famous-Roasters/frontend/commit/be41dca6f594597f45511e6de60be0bb805c9f8a))
+
 # [1.13.0](https://github.com/Famous-Roasters/frontend/compare/v1.12.0...v1.13.0) (2022-07-05)